Slave mode
If Slave Mode is enabled, JoeyNMT loads the model and then waits for standard input. When input is supplied, its translation is printed to standard output. Then, JoeyNMT WAITS FOR MORE INPUT instead of exiting. It will only exit if the process is killed or a signal is issued through standard input.